# Optimized demultiplexing