… woltka/tests/data/align/blastn/mux.b6o.xz
… woltka/tests/data/output/blastn.species.tsv