Benchmarks on 9,921 lines belonging to five samples:

- Before: 5.03 ms ± 53.2 µs
- After: 2.27 ms ± 29 µs

Also tested alternative solutions (e.g., using `defaultdict` instead of `dict`). The current is the best.